(PC-40911)[PRO] refactor: `SiretInputForm`.
