+++ pro/src/components/SignupJourneyForm/Offerer/Offerer.module.scss
@@ -9,10 +9,8 @@
 
 .subtitle {
   @include fonts.title3;
-}
 
-.input-siret {
-  margin-top: var(--size-spacing-xxl);
+  margin-bottom: var(--size-spacing-xxl);
 }
 
 .main-heading {
